Extend backup/restore tests
This patch enables the restore test case and extends both the backup and
restore tests so that a key is generated, data is encrypted for the key
before the backup is created and then decrypted after the backup has
been restored.  It also adds a test to make sure that the factory reset
clears the generated key.

Fixes: https://github.com/Nitrokey/nethsm-sdk-py/issues/48