Add certifi dependency (#47)
* Move version into nethsm/__init__.py

To make it possible to use `pip install .`, we have to change
`__init__.py` so that it is possible to determine the version with
static analysis.

See this issue for more context:
    https://github.com/Nitrokey/nethsm-sdk-py/issues/33

* ci: Add clean-install job

This patch adds a clean-install job to the CI that makes sure that
installing the package with pip and using it works, i. e. we did not
forget to specify a required dependency.

* Add missing certifi dependency

This patch adds certifi as a required dependency.  Previously, we did
not notice it was missing as it was installed together with flit in the
CI and development setup.

Fixes: https://github.com/Nitrokey/nethsm-sdk-py/issues/33