Make authentication optional
Not all endpoints require authentication.  Previously, it was already
possible to pass None for the username and password although it was not
indicated by the type annotations.  This patch adds the Authentication
data class to make it easier to pass or omit these arguments and adjusts
the type annotations accordingly.

It also adds a test to make sure that accessing the system state works
without authentication.

Fixes: https://github.com/Nitrokey/nethsm-sdk-py/issues/68