Mark exception handlers as NoReturn
The _handle_exception and _handle_api_exception functions always raise
an exception.  Previously, these were annotated as returning None so
that mypy could not determine that they never return normally.  We had
to add assertions to indicate that certain codepaths are never used.
With this patch, we return NoReturn instead so that the assertions are
no longer necessary.

Fixes: https://github.com/Nitrokey/nethsm-sdk-py/issues/59